Created the endpoint and test to listall users
