+++ src/main/java/br/com/caelum/carangobom/user/UserController.java
@@ -8,6 +8,9 @@ import org.springframework.web.util.UriComponentsBuilder;
 import javax.transaction.Transactional;
 import javax.validation.Valid;
 import java.net.URI;
+import java.util.Collections;
+import java.util.List;
+import java.util.Optional;
 
 @RestController
 public class UserController {
@@ -21,9 +24,15 @@ public class UserController {
     }
 
 
+    @GetMapping("/users")
+    public List<UserDTO> list_all_users() {
+        List<User> users = userRepository.findAll();
+        return UserDTO.convert(users);
+    }
+
     @PostMapping("/users")
     @Transactional
-    public ResponseEntity<UserDTO> create(@RequestBody @Valid UserForm userForm, UriComponentsBuilder uriBuilder){
+    public ResponseEntity<UserDTO> create(@RequestBody @Valid UserForm userForm, UriComponentsBuilder uriBuilder) {
         User user = userForm.convert();
 
         userRepository.save(user);
