+++ src/main/java/br/com/caelum/carangobom/user/UserRepository.java
@@ -3,5 +3,7 @@ package br.com.caelum.carangobom.user;
 import br.com.caelum.carangobom.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.List;
+
 public interface UserRepository extends JpaRepository<User, Long> {
 }
