--- src/test/java/br/com/caelum/carangobom/user/SpringUserControllerTest.java
@@ -1,66 +1,0 @@
-package br.com.caelum.carangobom.user;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.MediaType;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.junit4.SpringRunner;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
-import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
-import java.net.URI;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-@AutoConfigureMockMvc
-@ActiveProfiles("test")
-public class SpringUserControllerTest {
-
-    @Autowired
-    private MockMvc mockMvc;
-
-    @Test
-    public void shouldCreateANewUser() throws Exception {
-        URI uri = new URI("/users");
-
-        String json = "{\"username\": \"username\", \"password\": \"newpassword\"}";
-
-
-        mockMvc.perform(MockMvcRequestBuilders
-                .post(uri)
-                .content(json)
-                .contentType(MediaType.APPLICATION_JSON))
-                .andExpect(MockMvcResultMatchers.status().is2xxSuccessful());
-    }
-
-    @Test
-    public void shouldReturnErrorWhenDoesNotHaveUsername() throws Exception {
-        URI uri = new URI("/users");
-
-        String json = "{\"username\": \"\", \"password\": \"newpassword\"}";
-
-
-        mockMvc.perform(MockMvcRequestBuilders
-                .post(uri)
-                .content(json)
-                .contentType(MediaType.APPLICATION_JSON))
-                .andExpect(MockMvcResultMatchers.status().is4xxClientError());
-    }
-
-    @Test
-    public void shouldReturnErrorWhenDoesNotHavePassword() throws Exception {
-        URI uri = new URI("/users");
-
-        String json = "{\"username\": \"username\", \"password\": \"\"}";
-
-        mockMvc.perform(MockMvcRequestBuilders
-                .post(uri)
-                .content(json)
-                .contentType(MediaType.APPLICATION_JSON))
-                .andExpect(MockMvcResultMatchers.status().is4xxClientError());
-    }
-
-}
